Do not throw error if model not found for indexing

+++ packages/component-elasticsearch/src/component-elasticsearch.service.ts
@@ -28,13 +28,15 @@ class ComponentElasticsearchService {
   getIndexName(modelName: string): string {
     const [modelConfig] = this.MongoEsIndexer.configs.filter(i=>i.model === modelName);
     if(!modelConfig) {
-      throw new Error('No model config found for model: ' + modelName);
+      // throw new Error('No model config found for model: ' + modelName);
+      return;
     }
     return modelConfig.indexName.split(this.MongoEsIndexer.indexPrefix)[1];
   }
 
   @MongoOplogHandler('insert')
   async indexOne({ collection: modelName, oplogData }) {
+    if(!this.getIndexName(modelName)) return;
     try {
       await this.MongoEsIndexer.indexOne(
         `${this.MongoEsIndexer.indexPrefix}${this.getIndexName(modelName)}`,
@@ -47,6 +49,7 @@ class ComponentElasticsearchService {
 
   @MongoOplogHandler('update')
   async updateOne({ collection: modelName, oplogData }) {
+    if(!this.getIndexName(modelName)) return;
     try {
       await this.MongoEsIndexer.updateOne(
         `${this.MongoEsIndexer.indexPrefix}${this.getIndexName(modelName)}`,
@@ -59,6 +62,7 @@ class ComponentElasticsearchService {
 
   @MongoOplogHandler('delete')
   deleteByIds({ collection: modelName, oplogData }): Promise<any> {
+    if(!this.getIndexName(modelName)) return;
     return this.MongoEsIndexer.deleteByIds(`${this.MongoEsIndexer.indexPrefix}${this.getIndexName(modelName)}`, [
       oplogData?.documentKey?._id,
     ]);
